feat: move connect device page to pages folder
