+++ src/pages/ConnectDevicePage/ConnectDevicePage.tsx
@@ -1,15 +1,15 @@
 import { useState } from 'react'
-import BreadcrumbBar from './BreadcrumbBar'
+import BreadcrumbBar from '../../components/BreadcrumbBar'
 import { Button as StatusButton, Tag, Text, Avatar, Checkbox } from '@status-im/components'
 import { Label, Separator, XStack, YStack } from 'tamagui'
-import PageWrapperShadow from './PageWrapperShadow'
-import NimbusLogo from './NimbusLogo'
-import Titles from './Titles'
-import NodeIcon from './NodeIcon'
-import ConnectIcon from './ConnectIcon'
-import PairIcon from './PairIcon'
-import CreateIcon from './CreateIcon'
-import LabelInputField from './LabelInputField'
+import PageWrapperShadow from '../../components/PageWrapperShadow'
+import NimbusLogo from '../../components/NimbusLogo'
+import Titles from '../../components/Titles'
+import NodeIcon from '../../components/NodeIcon'
+import ConnectIcon from '../../components/ConnectIcon'
+import PairIcon from '../../components/PairIcon'
+import CreateIcon from '../../components/CreateIcon'
+import LabelInputField from '../../components/LabelInputField'
 
 const ConnectDevicePage = () => {
   const [autoConnectChecked, setAutoConnectChecked] = useState(false)
